+++ asurity-project-frontend/src/components/App.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {LoopCircleLoading} from 'react-loadingg'; 
 import '../styling/App.css';
 import ContactGrid from './ContactGrid';
-import ContactFormNew from './ContactFormNew';
+import ContactForm from './ContactForm';
 import {API} from '../Constants';
 import {Contact, NewContact, UsState, ContactMethod, ContactFrequency}from '../Interfaces';
 
@@ -21,13 +21,13 @@ interface State {
 }
 
 type View = 'contact grid' | 'contact form' | 'edit form'
+
 class App extends React.Component<Props, State> {
   constructor (props: Props) {
     super(props);
     this.state = { 
       view: 'contact grid',
       isLoading: false,
-      //error: null,
       contacts: [],
       stateSelections: [],
       contactMethods: [],
@@ -36,17 +36,21 @@ class App extends React.Component<Props, State> {
     };
   }
 
-  updateStateName (contact: Contact|NewContact, style: 'abbreviation'|'fullName'): void {
-    let stateName: string|undefined;
+  async componentDidMount () {
+    //loading
+    this.setState({isLoading: true})
 
-    if (style === 'abbreviation') {
-      stateName = this.state.stateSelections.find((stateData:UsState) => stateData.abbreviation === contact.state)?.fullName;
-    }
-    if (style === 'fullName') {
-      stateName = this.state.stateSelections.find((stateData:UsState) => stateData.fullName === contact.state)?.abbreviation;
-    }
-    
-    contact.state = stateName!;
+    //fetch dropdown selection data
+    fetch(API + "contactMethods")
+    .then(response => response.json())
+    .then(data => this.setState({contactMethods: data}));
+
+    fetch(API + "contactFrequencies")
+    .then(response => response.json())
+    .then(data => this.setState({contactFrequencies: data}));
+
+    //fetch contacts and US states
+    this.getStateInfoAndContacts();
   }
 
   //view handlers
@@ -68,7 +72,20 @@ class App extends React.Component<Props, State> {
     }
   } 
 
- //not sure how to get around this TypeScript Index Signature issue, need to investigate further
+  updateStateName (contact: Contact|NewContact, style: 'abbreviation'|'fullName'): void {
+    let stateName: string|undefined;
+
+    if (style === 'abbreviation') {
+      stateName = this.state.stateSelections.find((stateData:UsState) => stateData.abbreviation === contact.state)?.fullName;
+    }
+    if (style === 'fullName') {
+      stateName = this.state.stateSelections.find((stateData:UsState) => stateData.fullName === contact.state)?.abbreviation;
+    }
+    
+    contact.state = stateName!;
+  }
+
+  //not sure how to get around this TypeScript Index Signature issue, need to investigate further
   sortContacts = (dataPoint:string, sort: 'asc'|'desc') => {
     let sortedList: Contact[] = [...this.state.contacts];
     if ( sort === 'asc') {
@@ -79,48 +96,6 @@ class App extends React.Component<Props, State> {
       sortedList.sort((a,b) => a[dataPoint] < b[dataPoint] ? 1 : b[dataPoint] < a[dataPoint] ? -1 : 0);
     }
     this.setState({contacts: sortedList})
-  }
-
-  deleteContact = (id: number): void => {
-    fetch(API + 'contacts/' + id, {
-      method: 'DELETE'
-    }).then(() => {
-      let updatedContactList = this.state.contacts.filter(contact => contact.id !== id);
-      this.setState({contacts: updatedContactList});
-    })
-  }
-
-  editContact = (contact: Contact): void => {
-    this.updateStateName(contact, 'fullName');
-
-    fetch(API + 'contacts/' + contact.id, {
-      method: 'PUT',
-      body: JSON.stringify(contact),
-      headers: { 'Content-Type': 'application/json' }
-    }).then(response => response.json())
-    .then(contact => {
-      let newState = [...this.state.contacts];
-      this.updateStateName(contact, 'abbreviation');
-      let contactIndex = newState.findIndex(e => e.id === contact.id);
-      newState[contactIndex] = contact;
-      this.setState({contacts: newState})
-    })
-  }
-
-  addContact = (contact: NewContact): void => {
-    this.updateStateName(contact, 'fullName');
-
-    fetch(API + 'contacts', {
-      method: 'POST',
-      body: JSON.stringify(contact),
-      headers: { 'Content-Type': 'application/json' }
-    }).then(response => response.json())
-    .then(contact => {
-      let newState = [...this.state.contacts];
-      this.updateStateName(contact, 'abbreviation');
-      newState.push(contact);
-      this.setState({contacts: newState })
-    })
   }
 
   //contact population and state name population
@@ -139,21 +114,46 @@ class App extends React.Component<Props, State> {
     });
   }
 
-  async componentDidMount () {
-    //loading
-    this.setState({isLoading: true})
+  addContact = (contact: NewContact): void => {
+    this.updateStateName(contact, 'fullName');
 
-    //fetch dropdown selection data
-    fetch(API + "contactMethods")
-    .then(response => response.json())
-    .then(data => this.setState({contactMethods: data}));
+    fetch(API + 'contacts', {
+      method: 'POST',
+      body: JSON.stringify(contact),
+      headers: {'Content-Type': 'application/json'}
+    }).then(response => response.json())
+    .then(contact => {
+      let newState = [...this.state.contacts];
+      this.updateStateName(contact, 'abbreviation');
+      newState.push(contact);
+      this.setState({contacts: newState})
+    })
+  }
 
-    fetch(API + "contactFrequencies")
-    .then(response => response.json())
-    .then(data => this.setState({contactFrequencies: data}));
+  editContact = (contact: Contact): void => {
+    this.updateStateName(contact, 'fullName');
 
-    //fetch contacts and US states
-    this.getStateInfoAndContacts();
+    fetch(API + 'contacts/' + contact.id, {
+      method: 'PUT',
+      body: JSON.stringify(contact),
+      headers: {'Content-Type': 'application/json'}
+    }).then(response => response.json())
+    .then(contact => {
+      let newState = [...this.state.contacts];
+      this.updateStateName(contact, 'abbreviation');
+      let contactIndex = newState.findIndex(e => e.id === contact.id);
+      newState[contactIndex] = contact;
+      this.setState({contacts: newState})
+    })
+  }
+
+  deleteContact = (id: number): void => {
+    fetch(API + 'contacts/' + id, {
+      method: 'DELETE'
+    }).then(() => {
+      let updatedContactList = this.state.contacts.filter(contact => contact.id !== id);
+      this.setState({contacts: updatedContactList});
+    })
   }
 
   render () {
@@ -181,7 +181,7 @@ class App extends React.Component<Props, State> {
       mainViewComponent = <LoopCircleLoading />;
       headerButton = <span onClick={this.newContactClickHandler} >New Contact</span>
     } else if (this.state.view === 'contact form'){
-      mainViewComponent = <ContactFormNew {...contactFormProps} />;
+      mainViewComponent = <ContactForm {...contactFormProps} />;
       headerButton = <span onClick={this.gridViewHandler}>All Contacts</span>
     } else {
       mainViewComponent = <ContactGrid {...gridProps}/>;
